Implement support for issue #27.
MarkingStoryValidator will collect the priorities
of all potential steps and check if there is more
than one potential step with any given priority.
Only if that is true, a mark will be generated.

JBehaveProject now publicly exposes the getValue
method to access the annotation key/values more
easily.

PotentialStep's toString() has been slightly modified
to include the priority, if it is not 0.